feat(zero): WebSocket close beacon
When Zero close is called we send a special close reason that is
interpreted by the view syncer.

We send this close reason on pagehide when persisted is false. This
websocket send goes through in unload when the tab is closed or reloaded.
Different browsers deals with navigation differently.

https://replicache.notion.site/Close-Beacon-Yes-or-no-1ab3bed895458060bf8ad19603a2979f

In the view syncer, when we get this clean close we delete the client
which means that it will no longer make queries desired by that client.