feat(zero): Change default TTL to 10 seconds
The default TTL used to be undefined (which was used for no TTL). This
would leave queries around for a long time and cause a lot of extra work
to keep them up to date. This change sets the default TTL to 10 seconds
which should be a good balance between keeping queries around long enough
to be useful and not keeping them around too long.