feat(zero-client): only track the `main` branch for IVM
No longer tracks the sync head. Only tracks `main` and forks and re-winds as needed.

- refresh
- pullEnd
- persist

will all call `forkToHead` to get the correct copy of the IVM sources.

`advance` is currently not yet called. It will be called in a future PR when we have the data required (expectedHash, newHash, diffs). ExperimentalWatch does not provided the two hashes we need so I'll be updating Replicache to call `advance` directly.