feat(zero-client): Expose some internal information
This adds a new `Inspector` class to the Zero client that can be used to
inspect the internal state of the client. This is useful for debugging
purposes.

The inspector is lazily loaded so that the client does not need to pay
the downloading cost if it is not used.

The inspector only exposes a few things at the moment:

- All the known clients in this IDB store. There can be multiple clients
  groups in the IDB instance and these groups can have multiple clients.
- The queries desired queries for a client and whether they are "got".
- The AST of the query and an SQL representation of the query.

```ts
const inspector = await z.inspect();
for (const query of await inspector.client.queries()) {
  console.log(query.id);
  console.log(query.got);
  console.log(query.sql);
}
```

This is a first step towards exposing more internal information.