fix(zql): Fix ref count for edit in applyChange
The edit branch was incomplete.

If we edit the value a non primary column, we transfer the ref count to
the new object.

If we edit the value of a primary column, we need to decrement the ref
count of the old object and increment the ref count of the new object.