Setup heartbeats for websockets
Adds logic so that every 3 seconds the server sends a ping frame, and checks for pong's coming back. If it doesn't get a pong back the server will terminate the connection.

Then also adds logic on the client to expect the pings, and if doesn't receive one for a while also terminate the connection there.