fix(z2s): handle self join, not exists
When we added name mapping the compiler got out of hand. This restructures the compiler so we are not re-mapping names all over the place as well as applies auto-aliasing to tables so we can do self joins.

E.g., before auto-aliasing we had:

```sql
SELECT (SELECT * FROM employee WHERE employee.reports_to = employee.id) as reports_to FROM employee;
```

which would return no results for the sub-select since the sub-query never references the outer-query.

Now we have:

```sql
SELECT (SELECT * FROM employee as employee_1 WHERE employee_1.reports_to = employee_0.id) as reports_to FROM employee as employee_0
```

`NOT EXISTS` forgot to apply the correlation. We pass the correlation to `NOT EXISTS` now.