fix(zql): pk constraint is a filter, not the constraint
The primary key constraint that we build from the connection's filters should not take the place of `req.constraint` everywhere.

The reason is that the primary key constraint must be ANDED with the fetch constraint. Example:

An album node could `fetch(constrating: artistID = 2)`. The connection may have a primary key filter on it `artistID = 1` but it would be wrong to return that artist since it was not `artistID = 2`.