chore(zero-client): Add active clients detection
This uses a combination of web locks and broadcast channel.

Each client sets up an exclusive lock for its name, and then uses a
broadcast channel to announce its presence.

Other clients listen for these announcements on the broadcast channel
and sets up a shared lock for the client's name. This way the shared
lock will be acquired when the client goes away.

In case web locks are not supported we use an in memory fallback which
at least allows multiple clients in the same global scope to work
together.