chore(zero): Remove closeConnection and close beacon
This deprecates the `closeConnection` protocol message and removes the
`closeConnection` beacon that was sent by the client when the client was
closed.

The `closeConnection` message was used to inactivate queries when
closing the client. We are going to use a different approach to
inactivate queries and this does not involve inactivating queries when
the client is closed.