fix(ttl): fix incorrect calculation of expired queries
There are two paths the check if queries are expired:

1. `hasExpiredQueries` in view-syncer.ts
2. `getInactiveQueries` in cvr.ts

(1) only returns queries that are expired for all clients.
(2) was returning queries if they were expired for 1 or more clients.

On reading, it appears that both paths should caluclate expiration the same way. The second path should not return a query as inactive if one of its client has not inactivated it.