feat(zero-client): Use admin password for inspector
Add support for admin password authentication in the inspector.

The server can now respond to any inspect rpc with authorized false. If
the client receives this, it will prompt the user for the admin password
and send a authenticate rpc request. When a successful auth response is
received, the client then retries the original inspect rpc.

On the server side, the inspector delegate keeps tracks of authenticated
client groups by keeping a set in memory.