fix(replicache/zero-client):
This fixes a few issues around close and destroy for the SQLite KV store
and the Expo wrapper for it.

The expo wrapper now makes sure it does not call close on an already
closed database.

The database manager manages multiple instances to ensure the database
behaves as a RWLock. It use multiple connections to do this. However,
there was a confusion who owned what and the code ended up trying to
finalize (cleanup) prepared statements that belonged to a closed db. The
database manager now manages all the cleanup and closing.

The instances managed by the database manager can be in two states; open
or closed. Since we keep the instances around until the database is
destroyed we need to keep track of this state so that we do not try to
finalize/close an already closed database.

Added more tests around closing and destroying the database instances.