chore(planner): ANDed exists are additive, not multiplicative
Thought I addressed this while addressing `OR`. Apparently not.

```ts
issue
  .whereExists('label')
  .whereExists('creator')
```

The costs should be added in this case as the same base row (those from issue) accumulate child rows in a pipeline rather than a cartesian product forming.

I.e., we do not get a `creator` for every `issue,label` pairing.

exists is limit 1 but so one could argue mult vs addition is the same here but that's not true since selectivity could cause a large scan, regardless of limit.