Bump Zero version to 0.24.3000.
This version was picked so that it is bigger than any of the prior
date-based version numbers being used as canaries.

We recently changed the way canary scripts are being numbered,
now they look like <major>.<minor>.<patch>-canary.<attempt> and
the release script increments the current m.m.v found in the code.

So we need to bump the base version to some semver higher than
previously released 0.24's.