fix(zero-cache): handle connections to a view-syncer that has shut down (#5101)
Handle a race condition in which an incoming connection is made after
the view-syncer has shutdown (but before the ServiceRunner has removed
it from its map).

Such connections used to be black-holed; now they are disconnected with
a `Rehome` protocol error, which the client handles by reconnecting
after a backoff.

User report:
*
https://discord.com/channels/830183651022471199/1433155873127272608/1433169650333782148