fix(zero-cache): do not hydrate queries if no clients initialized their connections
The problem is that custom queries hit the API server.

Some users _require_ that all of their users are logged in.

If the custom query endpoint does not support logged out users, it will throw and tear down the view syncer.
This disconnects everyone, they try reconnecting again, same problem happens.

This loop will only stop happening if a `initConnection` message happens to be processed before we try transforming a custom query.