fix(zero-cache): process pending messages on closed websocket (#5110)
In the change-streamer <-> replicator connection, the websocket may be
closed by the change-streamer due to the subscriber being unresponsive
(e.g. cpu blocked on an expensive SQLite statement, such as index
creation).

When this happens, process the pending messages on the subscriber side
even if the websocket is closed. This increases the chance of making
forward progress by committing the transaction containing the expensive
statement.

Context:
https://www.notion.so/replicache/Index-change-backlog-29c3bed8954580bfac64de81aee02db4?source=copy_link#29c3bed89545809a80aed04f27a5d5d0