fix(zero-cache): catch errors from async/pipelined db calls (#5163)
A peculiar case of PG serialization errors in the CVR logic (which is
still being root-caused) revealed some async / pipelined db calls for
which exceptions are not being handled, thereby crashing the server.


https://discord.com/channels/830183651022471199/1437863069375135795/1437879816933212346

Fix those call sites to properly handle exception.

Investigation of why the PG error is happening more often is ongoing.