fix(zero-cache): expliclty set isolation level in CVR transactions (#5165)
The database calls in the CVR code involve a certain amount of
asynchrony that rely on the fact that the transactions are performed
with the default `READ COMMITTED` isolation level. This assumption can
be broken if the database has a different isolation level setting. To
protect against this, explicitly set the required isolation level.

User report:
*
https://discord.com/channels/830183651022471199/1437863069375135795/1437945258683138310