feat(zero-server): add simplified handleMutateRequest API
Deprecate handleMutationRequest and replace with a simplified version.

It is not robust to do asynchronous work in handleMutationRequest
after the transaction boundary (no guarantee it gets done). The
correct way to send notifications is w/ a transactional outbox
pattern, and that's what we should encourage.

If people do want to do a quick hack to send notifications this way,
then the current pattern of appending a task to an array is actually
preferrable since they can all be done in parallel after the the tx
completes.