chore(zero): Add test for protocol version URL hash
If the URL search params change, we need to bump the protocol version.

This adds a test that will fail if the URL search params change,
reminding us to bump the protocol version.

And updates PROTOCOL_VERSION to 44 to account for the new profileID
param added to the connection URL.