fix(zero-cache): added better handling of permissions (#5341)
Permissions are still half-broken currently, for some of the older
scripts.

This adds:
- Warnings when running `zero-cache-dev` to remove the `-p` option or
`ZERO_SCHEMA_PATH` env (see
https://discord.com/channels/830183651022471199/1445578476277993532/1449049784559210678).
- **Fails to run when both `ZERO_MUTATE_URL` and `ZERO_QUERY_URL` and
`-p` are defined, since permissions in this case should not be used.**
- Deprecation warnings printed to console when permissions are defined.
- Adds a test to make sure that `loadSchemaAndPermissions` doesn't fail
when permissions aren't defined (still used across `analyze-query`,
`transform-query`, and `ast-to-zql` to pull the schema and apply
permissions on top of queries.