chore(zero-cache): traverse changeLog entries chronologically (#5372)
Revival of #4956

Change the schema of the replica's `changeLog` table such that changes
are ordered (and traversed) chronologically, i.e. in the order in which
they were received from upstream.

This reduces the frequency in which manual deletes must be pushed into
IVM in order to avoid unique constraint violations
(https://github.com/rocicorp/mono/pull/5030).

Context of what https://github.com/rocicorp/mono/pull/5030 fixes:
*
https://discord.com/channels/830183651022471199/1422347526283071580/1422638434740207783


A chronological changeLog also makes IVM pushes more intuitive to reason
about, in that changes are pushed in the same order as they appear in
the replication stream. However, the `changeLog` still maintains a
unique index on the `<table, rowKey>` columns such that only the _last_
change to a row is kept, maintaining the O(num-rows) bound on the size
of the log and performance benefit of removing redundant ivm pushes.

### Compatibility Note

In order for new replica files to remain usable by older zero-cache
code, a new `changeLog2` table is created and used so that older code
can continue using the old `changeLog` table.